+++ zc-aliyun-spring-boot-starter/src/main/java/com/zhengcheng/aliyun/mq/annotation/Event.java
@@ -1,6 +1,9 @@
 package com.zhengcheng.aliyun.mq.annotation;
 
 
+import org.springframework.core.annotation.AliasFor;
+import org.springframework.stereotype.Component;
+
 import java.lang.annotation.*;
 
 /**
@@ -8,10 +11,23 @@ import java.lang.annotation.*;
  *
  * @author :    quansheng.zhang
  * @date :    2019/8/12 22:41
+ * @see Component
+ * @since 4.1.0
  */
 @Target(ElementType.TYPE)
 @Retention(RetentionPolicy.RUNTIME)
 @Documented
+@Component
 public @interface Event {
-    String[] value();
+
+    /**
+     * The value may indicate a suggestion for a logical component name,
+     * to be turned into a Spring bean in case of an autodetected component.
+     *
+     * @return the suggested component name, if any (or empty String otherwise)
+     */
+    @AliasFor(annotation = Component.class)
+    String value() default "";
+
+    String[] tags() default {};
 }